wip: LogInput
